users model, routes and controllers

+++ back/src/controllers/employees/deleteEmployee.js
@@ -3,12 +3,13 @@ const { employees } = require("../../db");
 const deleteEmployees = async (req, res) => {
 	try {
 		const { id } = req.params;
-		await employees.destroy({
-			where: {
-				id,
-			},
-		});
-		res.status(200).json({ success: true });
+		const find = employees.findOne({ where: { id } });
+		if (find) {
+			await employees.destroy({
+				where: {id}
+			});
+			res.status(200).json({ success: true });
+		} else throw new Error("Could not find the employee");
 	} catch (err) {
 		res.status(400).json({ message: err.message });
 	}
